Move "global_worker_pool" to "src/util"
