+++ src/style/style.js
@@ -20,7 +20,7 @@ const QueryFeatures = require('../source/query_features');
 const SourceCache = require('../source/source_cache');
 const styleSpec = require('../style-spec/reference/latest');
 const MapboxGLFunction = require('../style-spec/function');
-const getWorkerPool = require('../global_worker_pool');
+const getWorkerPool = require('../util/global_worker_pool');
 const deref = require('../style-spec/deref');
 const diff = require('../style-spec/diff');
 const rtlTextPlugin = require('../source/rtl_text_plugin');
